pyproject.toml
[tool.poetry]
name = "awt-quant"
version = "0.1.0"
description = "Advanced Quantitative Forecasting with SPDE, GARCH, and LLMs"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "awt_quant"}]

# Define multiple sources CORRECTLY using Poetry's expected format
[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.3"
numpy = "^1.26.0"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
yfinance = "^0.2.31"
torch = { version = "2.6.0+cu118", source = "pytorch_cuda" }
torchvision = { version = "0.21.0+cu118", source = "pytorch_cuda" }
torchaudio = { version = "2.6.0+cu118", source = "pytorch_cuda" }



# Define extra dependencies for CPU and CUDA
lag-llama = {git = "https://github.com/time-series-foundation-models/lag-llama.git", rev = "update-gluonts"}
nixtlats = "^0.5.2"
autots = "^0.6.20"
python-dotenv = "^1.0.1"
[tool.poetry.extras]
cpu = ["torch", "torchvision", "torchaudio"]
cuda = ["torch", "torchvision", "torchaudio"]

[tool.poetry.group.dev.dependencies]
sphinx = "*"
sphinx-rtd-theme = "*"
pytest = "*"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
sphinx-rtd-theme = "^3.0.2"

[tool.poetry.scripts]
download-lag-llama = "scripts.download_model:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
